refactor(#792): split manifest_tests into base + NAM/enum/validation suite (sweep)
manifest_tests.rs 660 -> 338, + manifest_more_tests.rs 328 (both under cap; off debt).
The parse() helper becomes pub(super), imported via super::tests. plugin-loader
tests (6 groups) pass, zero warnings.
